move routes stuff to its own component
